clang-tidy

+++ core.mod/src/WGDefault.cc
@@ -1,8 +1,8 @@
 #include "WGDefault.h"
 
 void WGDefault::genChunk(Swan::WorldPlane &plane, Swan::Chunk &chunk) {
-	int height = perlin_.octaveNoise0_1(chunk.pos_.x / (64.0 / 16), 8) * 10;
-	int depth = perlin_.octaveNoise0_1(chunk.pos_.x / (64.0 / 9), 5) * 10 + 10;
+	int height = (int)(perlin_.octaveNoise0_1(chunk.pos_.x / (64.0 / 16), 8) * 10);
+	int depth = (int)(perlin_.octaveNoise0_1(chunk.pos_.x / (64.0 / 9), 5) * 10 + 10);
 
 	for (int cx = 0; cx < Swan::CHUNK_WIDTH; ++cx) {
 		for (int cy = 0; cy < Swan::CHUNK_HEIGHT; ++cy) {
